The iSEEDESeq2Results class

The iSEEDESeq2Results class is used to provide an common interface to differential expression results produced by the limma package. It provides methods to access common differential expression statistics (e.g., log2 fold-change, p-value, log2 average abundance).

Details

This class inherits all its slots directly from its parent class DataFrame.

Constructor

iSEEDESeq2Results(data, row.names = rownames(data)) creates an instance of a iSEEDESeq2Results class, with:

data

A data.frame produced by DESeq2::results() or DESeq2::lfcShrink().

row.names

The character vector of rownames for the SummarizedExperiment object in which the object is to be embedded. Must be a superset of rownames(data).

Supported methods

embedContrastResults(x, se, name, ...) embeds x in se under the identifier name. See embedContrastResults() for more details.

pValue(x) returns the vector of raw p-values.

log2FoldChange(x) returns the vector of log2-fold-change values.

averageLog2(x) returns the vector of average log2-expression values.

Author

Kevin Rue-Albrecht

Examples

library(DESeq2)

##
# From DESeq2::DESeq() ----
##

cnts <- matrix(rnbinom(n = 1000, mu = 100, size = 1 / 0.5), ncol = 10)
rownames(cnts) <- paste("Gene", 1:100)
cond <- factor(rep(1:2, each = 5))

# object construction
dds <- DESeqDataSetFromMatrix(cnts, DataFrame(cond), ~cond)
#> converting counts to integer mode

# standard analysis
dds <- DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.
#> final dispersion estimates
#> fitting model and testing
res <- results(dds)
head(res)
#> log2 fold change (MLE): cond 2 vs 1 
#> Wald test p-value: cond 2 vs 1 
#> DataFrame with 6 rows and 6 columns
#>         baseMean log2FoldChange     lfcSE      stat     pvalue      padj
#>        <numeric>      <numeric> <numeric> <numeric>  <numeric> <numeric>
#> Gene 1   46.8724       0.704892  0.645978  1.091202 0.27518418  0.749724
#> Gene 2   76.8912       0.296629  0.571491  0.519045 0.60372964  0.848172
#> Gene 3   99.0916       0.184074  0.547812  0.336016 0.73685878  0.870132
#> Gene 4  138.1972       0.976872  0.566764  1.723598 0.08478049  0.483333
#> Gene 5   83.8354       1.103274  0.623453  1.769620 0.07679052  0.483333
#> Gene 6   45.2272      -1.682580  0.631784 -2.663220 0.00773969  0.128995

##
# iSEEDESeq2Results ----
##

# Embed the DESeq2 results in the SummarizedExperiment object
dds <- embedContrastResults(res, dds, name = "DESeq2")

##
# Access ----
##

contrastResultsNames(dds)
#> [1] "DESeq2"
contrastResults(dds)
#> DataFrame with 100 rows and 1 column
#>                       DESeq2
#>          <iSEEDESeq2Results>
#> Gene 1   <iSEEDESeq2Results>
#> Gene 2   <iSEEDESeq2Results>
#> Gene 3   <iSEEDESeq2Results>
#> Gene 4   <iSEEDESeq2Results>
#> Gene 5   <iSEEDESeq2Results>
#> ...                      ...
#> Gene 96  <iSEEDESeq2Results>
#> Gene 97  <iSEEDESeq2Results>
#> Gene 98  <iSEEDESeq2Results>
#> Gene 99  <iSEEDESeq2Results>
#> Gene 100 <iSEEDESeq2Results>
contrastResults(dds, "DESeq2")
#> iSEEDESeq2Results with 100 rows and 6 columns
#>           baseMean log2FoldChange     lfcSE      stat     pvalue      padj
#>          <numeric>      <numeric> <numeric> <numeric>  <numeric> <numeric>
#> Gene 1     46.8724       0.704892  0.645978  1.091202  0.2751842  0.749724
#> Gene 2     76.8912       0.296629  0.571491  0.519045  0.6037296  0.848172
#> Gene 3     99.0916       0.184074  0.547812  0.336016  0.7368588  0.870132
#> Gene 4    138.1972       0.976872  0.566764  1.723598  0.0847805  0.483333
#> Gene 5     83.8354       1.103274  0.623453  1.769620  0.0767905  0.483333
#> ...            ...            ...       ...       ...        ...       ...
#> Gene 96    91.6270      -1.079595  0.649876 -1.661233 0.09666669  0.483333
#> Gene 97    90.8868      -0.172300  0.451160 -0.381905 0.70253183  0.870132
#> Gene 98    98.4329       0.143400  0.654440  0.219118 0.82655808  0.918398
#> Gene 99    82.9777       0.380758  0.719308  0.529339 0.59656998  0.848172
#> Gene 100  116.6769       1.592422  0.595515  2.674026 0.00749467  0.128995

head(pValue(contrastResults(dds, "DESeq2")))
#>      Gene 1      Gene 2      Gene 3      Gene 4      Gene 5      Gene 6 
#> 0.275184177 0.603729638 0.736858778 0.084780492 0.076790525 0.007739686 
head(log2FoldChange(contrastResults(dds, "DESeq2")))
#>     Gene 1     Gene 2     Gene 3     Gene 4     Gene 5     Gene 6 
#>  0.7048918  0.2966295  0.1840736  0.9768724  1.1032744 -1.6825796 
head(averageLog2(contrastResults(dds, "DESeq2")))
#>   Gene 1   Gene 2   Gene 3   Gene 4   Gene 5   Gene 6 
#> 5.550667 6.264747 6.630690 7.110585 6.389487 5.499118