The iSEEDESeq2Results class

The iSEEDESeq2Results class is used to provide an common interface to differential expression results produced by the limma package. It provides methods to access common differential expression statistics (e.g., log2 fold-change, p-value, log2 average abundance).

Details

This class inherits all its slots directly from its parent class DataFrame.

Constructor

iSEEDESeq2Results(data, row.names = rownames(data)) creates an instance of a iSEEDESeq2Results class, with:

data

A data.frame produced by DESeq2::results() or DESeq2::lfcShrink().

row.names

The character vector of rownames for the SummarizedExperiment object in which the object is to be embedded. Must be a superset of rownames(data).

Supported methods

pvalue(x) returns the vector of raw p-values.

log2foldchange(x) returns the vector of log2-fold-change values.

log2average(x) returns the vector of average log2-expression values.

Author

Kevin Rue-Albrecht

Examples

library(DESeq2)
#> Loading required package: S4Vectors
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomicRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians

##
# From DESeq2::DESeq() ----
##

cnts <- matrix(rnbinom(n=1000, mu=100, size=1/0.5), ncol=10)
rownames(cnts) <- paste("Gene",1:100)
cond <- factor(rep(1:2, each=5))

# object construction
dds <- DESeqDataSetFromMatrix(cnts, DataFrame(cond), ~ cond)
#> converting counts to integer mode

# standard analysis
dds <- DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.
#> final dispersion estimates
#> fitting model and testing
res <- results(dds)
head(res)
#> log2 fold change (MLE): cond 2 vs 1 
#> Wald test p-value: cond 2 vs 1 
#> DataFrame with 6 rows and 6 columns
#>         baseMean log2FoldChange     lfcSE      stat    pvalue      padj
#>        <numeric>      <numeric> <numeric> <numeric> <numeric> <numeric>
#> Gene 1   86.4879     -0.2462765  0.477683 -0.515565  0.606158  0.863718
#> Gene 2   87.5401     -0.8721292  0.580242 -1.503044  0.132828  0.646835
#> Gene 3   73.5558      0.2461810  0.525569  0.468409  0.639492  0.863718
#> Gene 4   79.2156      0.6005590  0.755196  0.795236  0.426476  0.863718
#> Gene 5  113.5310      0.0741951  0.716069  0.103615  0.917475  0.954340
#> Gene 6   96.9846     -0.3075691  0.664688 -0.462727  0.643560  0.863718

##
# iSEEDESeq2Results ----
##

# Package the results in a iSEEDESeq2Results object
iseede_table <- iSEEDESeq2Results(res, row.names=rownames(dds))

# Store the iSEEDESeq2Results object in the SummarizedExperiment rowData
rowData(dds)[["iSEEde"]] <- DataFrame(DESeq2=I(iseede_table))

dds
#> class: DESeqDataSet 
#> dim: 100 10 
#> metadata(1): version
#> assays(4): counts mu H cooks
#> rownames(100): Gene 1 Gene 2 ... Gene 99 Gene 100
#> rowData names(23): baseMean baseVar ... maxCooks iSEEde
#> colnames: NULL
#> colData names(2): cond sizeFactor

##
# Methods ----
##

head(pvalue(iseede_table))
#>    Gene 1    Gene 2    Gene 3    Gene 4    Gene 5    Gene 6 
#> 0.6061583 0.1328277 0.6394923 0.4264764 0.9174753 0.6435601 
head(log2foldchange(iseede_table))
#>      Gene 1      Gene 2      Gene 3      Gene 4      Gene 5      Gene 6 
#> -0.24627653 -0.87212924  0.24618103  0.60055899  0.07419514 -0.30756910 
head(log2average(iseede_table))
#>   Gene 1   Gene 2   Gene 3   Gene 4   Gene 5   Gene 6 
#> 6.434426 6.451872 6.200768 6.307713 6.826942 6.599684