The iSEEDESeq2Results class

The iSEEDESeq2Results class is used to provide a common interface to differential expression results produced by the DESeq2 package. It provides methods to access common differential expression statistics (e.g., log2 fold-change, p-value, log2 average abundance).

Details

This class inherits all its slots directly from its parent class DataFrame.

Constructor

iSEEDESeq2Results(data, row.names = rownames(data)) creates an instance of a iSEEDESeq2Results class, with:

data

A data.frame produced by DESeq2::results() or DESeq2::lfcShrink().

row.names

The character vector of rownames for the SummarizedExperiment object in which the object is to be embedded. Must be a superset of rownames(data).

Supported methods

embedContrastResults(x, se, name, ...) embeds x in se under the identifier name. See embedContrastResults() for more details.

pValue(x) returns the vector of raw p-values.

log2FoldChange(x) returns the vector of log2-fold-change values.

averageLog2(x) returns the vector of average log2-expression values.

Author

Kevin Rue-Albrecht

Examples

library(DESeq2)

##
# From DESeq2::DESeq() ----
##

cnts <- matrix(rnbinom(n = 1000, mu = 100, size = 1 / 0.5), ncol = 10)
rownames(cnts) <- paste("Gene", 1:100)
cond <- factor(rep(1:2, each = 5))

# object construction
dds <- DESeqDataSetFromMatrix(cnts, DataFrame(cond), ~cond)
#> converting counts to integer mode

# standard analysis
dds <- DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
res <- results(dds)
head(res)
#> log2 fold change (MLE): cond 2 vs 1 
#> Wald test p-value: cond 2 vs 1 
#> DataFrame with 6 rows and 6 columns
#>         baseMean log2FoldChange     lfcSE       stat    pvalue      padj
#>        <numeric>      <numeric> <numeric>  <numeric> <numeric> <numeric>
#> Gene 1   96.0150      0.7359010  0.670382  1.0977337  0.272321  0.861275
#> Gene 2   93.6464      0.7092930  0.577425  1.2283718  0.219307  0.861275
#> Gene 3   79.4442     -0.9308195  0.614376 -1.5150646  0.129756  0.830290
#> Gene 4   81.5428      0.7742114  0.526129  1.4715243  0.141149  0.830290
#> Gene 5   94.5390      0.0370881  0.441604  0.0839851  0.933068  0.993949
#> Gene 6   99.7855      0.6908079  0.521485  1.3246949  0.185272  0.861275

##
# iSEEDESeq2Results ----
##

# Embed the DESeq2 results in the SummarizedExperiment object
dds <- embedContrastResults(res, dds, name = "DESeq2")

##
# Access ----
##

contrastResultsNames(dds)
#> [1] "DESeq2"
contrastResults(dds)
#> DataFrame with 100 rows and 1 column
#>                       DESeq2
#>          <iSEEDESeq2Results>
#> Gene 1   <iSEEDESeq2Results>
#> Gene 2   <iSEEDESeq2Results>
#> Gene 3   <iSEEDESeq2Results>
#> Gene 4   <iSEEDESeq2Results>
#> Gene 5   <iSEEDESeq2Results>
#> ...                      ...
#> Gene 96  <iSEEDESeq2Results>
#> Gene 97  <iSEEDESeq2Results>
#> Gene 98  <iSEEDESeq2Results>
#> Gene 99  <iSEEDESeq2Results>
#> Gene 100 <iSEEDESeq2Results>
contrastResults(dds, "DESeq2")
#> iSEEDESeq2Results with 100 rows and 6 columns
#>           baseMean log2FoldChange     lfcSE        stat    pvalue      padj
#>          <numeric>      <numeric> <numeric>   <numeric> <numeric> <numeric>
#> Gene 1     96.0150      0.7359010  0.670382   1.0977337  0.272321  0.861275
#> Gene 2     93.6464      0.7092930  0.577425   1.2283718  0.219307  0.861275
#> Gene 3     79.4442     -0.9308195  0.614376  -1.5150646  0.129756  0.830290
#> Gene 4     81.5428      0.7742114  0.526129   1.4715243  0.141149  0.830290
#> Gene 5     94.5390      0.0370881  0.441604   0.0839851  0.933068  0.993949
#> ...            ...            ...       ...         ...       ...       ...
#> Gene 96   125.1409    -1.16780721  0.590675 -1.97707114 0.0480336  0.655623
#> Gene 97   101.4181    -0.38327536  0.456512 -0.83957304 0.4011478  0.909385
#> Gene 98   121.8751    -1.36699464  0.627985 -2.17679356 0.0294960  0.655623
#> Gene 99    97.5391    -0.00446559  0.722937 -0.00617701 0.9950715  0.995071
#> Gene 100   59.8305     0.24176532  0.810898  0.29814534 0.7655922  0.982197

head(pValue(contrastResults(dds, "DESeq2")))
#>    Gene 1    Gene 2    Gene 3    Gene 4    Gene 5    Gene 6 
#> 0.2723208 0.2193074 0.1297561 0.1411494 0.9330683 0.1852724 
head(log2FoldChange(contrastResults(dds, "DESeq2")))
#>      Gene 1      Gene 2      Gene 3      Gene 4      Gene 5      Gene 6 
#>  0.73590098  0.70929299 -0.93081951  0.77421136  0.03708812  0.69080787 
head(averageLog2(contrastResults(dds, "DESeq2")))
#>   Gene 1   Gene 2   Gene 3   Gene 4   Gene 5   Gene 6 
#> 6.585188 6.549151 6.311870 6.349486 6.562837 6.640758