The iSEEDESeq2Results class

The iSEEDESeq2Results class is used to provide an common interface to differential expression results produced by the limma package. It provides methods to access common differential expression statistics (e.g., log2 fold-change, p-value, log2 average abundance).

Details

This class inherits all its slots directly from its parent class DataFrame.

Constructor

iSEEDESeq2Results(data, row.names = rownames(data)) creates an instance of a iSEEDESeq2Results class, with:

data

A data.frame produced by DESeq2::results() or DESeq2::lfcShrink().

row.names

The character vector of rownames for the SummarizedExperiment object in which the object is to be embedded. Must be a superset of rownames(data).

Supported methods

embedContrastResults(x, se, name, ...) embeds x in the column name of rowData(se)[["iSEEde"]].

pValue(x) returns the vector of raw p-values.

log2FoldChange(x) returns the vector of log2-fold-change values.

averageLog2(x) returns the vector of average log2-expression values.

Author

Kevin Rue-Albrecht

Examples

library(DESeq2)

##
# From DESeq2::DESeq() ----
##

cnts <- matrix(rnbinom(n = 1000, mu = 100, size = 1 / 0.5), ncol = 10)
rownames(cnts) <- paste("Gene", 1:100)
cond <- factor(rep(1:2, each = 5))

# object construction
dds <- DESeqDataSetFromMatrix(cnts, DataFrame(cond), ~cond)
#> converting counts to integer mode

# standard analysis
dds <- DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.
#> final dispersion estimates
#> fitting model and testing
res <- results(dds)
head(res)
#> log2 fold change (MLE): cond 2 vs 1 
#> Wald test p-value: cond 2 vs 1 
#> DataFrame with 6 rows and 6 columns
#>         baseMean log2FoldChange     lfcSE       stat    pvalue      padj
#>        <numeric>      <numeric> <numeric>  <numeric> <numeric> <numeric>
#> Gene 1  108.7035      0.4200876  0.611991  0.6864274  0.492444  0.962144
#> Gene 2   83.5636     -0.7723602  0.555491 -1.3904103  0.164404  0.962144
#> Gene 3   76.0398      0.2711216  0.564716  0.4801026  0.631154  0.962144
#> Gene 4   93.9459      0.0488246  0.591470  0.0825479  0.934211  0.963104
#> Gene 5  112.9068     -0.4386889  0.518840 -0.8455187  0.397821  0.962144
#> Gene 6   81.0921      0.5492644  0.620680  0.8849404  0.376189  0.962144

##
# iSEEDESeq2Results ----
##

# Package the results in a iSEEDESeq2Results object
iseede_table <- iSEEDESeq2Results(res, row.names = rownames(dds))

# Store the iSEEDESeq2Results object in the SummarizedExperiment rowData
rowData(dds)[["iSEEde"]] <- DataFrame(DESeq2 = I(iseede_table))

dds
#> class: DESeqDataSet 
#> dim: 100 10 
#> metadata(1): version
#> assays(4): counts mu H cooks
#> rownames(100): Gene 1 Gene 2 ... Gene 99 Gene 100
#> rowData names(23): baseMean baseVar ... maxCooks iSEEde
#> colnames: NULL
#> colData names(2): cond sizeFactor

##
# Methods ----
##

head(pValue(iseede_table))
#>    Gene 1    Gene 2    Gene 3    Gene 4    Gene 5    Gene 6 
#> 0.4924436 0.1644043 0.6311545 0.9342110 0.3978213 0.3761888 
head(log2FoldChange(iseede_table))
#>      Gene 1      Gene 2      Gene 3      Gene 4      Gene 5      Gene 6 
#>  0.42008756 -0.77236018  0.27112157  0.04882458 -0.43868895  0.54926442 
head(averageLog2(iseede_table))
#>   Gene 1   Gene 2   Gene 3   Gene 4   Gene 5   Gene 6 
#> 6.764255 6.384803 6.248682 6.553758 6.818989 6.341489