The iSEELimmaResults class

The iSEELimmaResults class is used to provide a common interface to differential expression results produced by the limma package. It provides methods to access common differential expression statistics (e.g., log fold-change, p-value, log2 average abundance).

Details

This class inherits all its slots directly from its parent class DataFrame.

Constructor

iSEELimmaResults(data, row.names = rownames(data)) creates an instance of a iSEELimmaResults class, with:

data

A data.frame produced by limma::topTable().

row.names

The character vector of rownames for the SummarizedExperiment object in which the object is to be embedded. Must be a superset of rownames(data).

Supported methods

embedContrastResults(x, se, name, class = "limma", ...) embeds x in se under the identifier name. See embedContrastResults() for more details.

pValue(x) returns the vector of raw p-values.

log2FoldChange(x) returns the vector of log2-fold-change values.

averageLog2(x) returns the vector of average log2-expression values.

Author

Kevin Rue-Albrecht

Examples

library(limma)
#> 
#> Attaching package: ‘limma’
#> The following object is masked from ‘package:DESeq2’:
#> 
#>     plotMA
#> The following object is masked from ‘package:BiocGenerics’:
#> 
#>     plotMA
library(SummarizedExperiment)

##
# From limma::lmFit() ----
##

sd <- 0.3 * sqrt(4 / rchisq(100, df = 4))
y <- matrix(rnorm(100 * 6, sd = sd), 100, 6)
rownames(y) <- paste("Gene", 1:100)
y[1:2, 4:6] <- y[1:2, 4:6] + 2
design <- cbind(Grp1 = 1, Grp2vs1 = c(0, 0, 0, 1, 1, 1))

fit <- lmFit(y, design)
fit <- eBayes(fit)
tt <- topTable(fit, coef = 2)
head(tt)
#>              logFC     AveExpr         t      P.Value    adj.P.Val         B
#> Gene 2   2.2131067  0.91543801 11.568403 2.697463e-06 0.0002697463  5.318800
#> Gene 1   1.8781911  0.85163875  7.271558 8.366161e-05 0.0041830807  1.657275
#> Gene 55  0.6235932 -0.06775434  3.092430 1.472702e-02 0.4909006606 -3.867864
#> Gene 77 -0.5702651 -0.22730316 -2.563958 3.328015e-02 0.6268675028 -4.703136
#> Gene 71  0.8479533  0.22931393  2.555530 3.372063e-02 0.6268675028 -4.716414
#> Gene 10 -0.7226727 -0.60014061 -2.402477 4.283301e-02 0.6268675028 -4.956655

##
# iSEELimmaResults ----
##

# Simulate the original SummarizedExperiment object
se <- SummarizedExperiment(assays = list(counts = y))

# Embed the Limma-Voom results in the SummarizedExperiment object
se <- embedContrastResults(tt, se, name = "Limma-Voom", class = "limma")

##
# Access ----
##

contrastResultsNames(se)
#> [1] "Limma-Voom"
contrastResults(se)
#> DataFrame with 100 rows and 1 column
#>                  Limma-Voom
#>          <iSEELimmaResults>
#> Gene 1   <iSEELimmaResults>
#> Gene 2   <iSEELimmaResults>
#> Gene 3   <iSEELimmaResults>
#> Gene 4   <iSEELimmaResults>
#> Gene 5   <iSEELimmaResults>
#> ...                     ...
#> Gene 96  <iSEELimmaResults>
#> Gene 97  <iSEELimmaResults>
#> Gene 98  <iSEELimmaResults>
#> Gene 99  <iSEELimmaResults>
#> Gene 100 <iSEELimmaResults>
contrastResults(se, "Limma-Voom")
#> iSEELimmaResults with 100 rows and 6 columns
#>              logFC   AveExpr         t     P.Value   adj.P.Val         B
#>          <numeric> <numeric> <numeric>   <numeric>   <numeric> <numeric>
#> Gene 1     1.87819  0.851639   7.27156 8.36616e-05 0.004183081   1.65727
#> Gene 2     2.21311  0.915438  11.56840 2.69746e-06 0.000269746   5.31880
#> Gene 3          NA        NA        NA          NA          NA        NA
#> Gene 4          NA        NA        NA          NA          NA        NA
#> Gene 5          NA        NA        NA          NA          NA        NA
#> ...            ...       ...       ...         ...         ...       ...
#> Gene 96         NA        NA        NA          NA          NA        NA
#> Gene 97         NA        NA        NA          NA          NA        NA
#> Gene 98         NA        NA        NA          NA          NA        NA
#> Gene 99         NA        NA        NA          NA          NA        NA
#> Gene 100        NA        NA        NA          NA          NA        NA

head(pValue(contrastResults(se, "Limma-Voom")))
#>       Gene 1       Gene 2       Gene 3       Gene 4       Gene 5       Gene 6 
#> 8.366161e-05 2.697463e-06           NA           NA           NA           NA 
head(log2FoldChange(contrastResults(se, "Limma-Voom")))
#>   Gene 1   Gene 2   Gene 3   Gene 4   Gene 5   Gene 6 
#> 1.878191 2.213107       NA       NA       NA       NA 
head(averageLog2(contrastResults(se, "Limma-Voom")))
#>    Gene 1    Gene 2    Gene 3    Gene 4    Gene 5    Gene 6 
#> 0.8516388 0.9154380        NA        NA        NA        NA