The iSEELimmaResults class

The iSEELimmaResults class is used to provide an common interface to differential expression results produced by the limma package. It provides methods to access common differential expression statistics (e.g., log fold-change, p-value, log2 average abundance).

# S4 method for iSEELimmaResults
embedResults(x, se, name, ...)

Details

This class inherits all its slots directly from its parent class DataFrame.

Constructor

iSEELimmaResults(data, row.names = rownames(data)) creates an instance of a iSEELimmaResults class, with:

data

A data.frame produced by limma::topTable().

row.names

The character vector of rownames for the SummarizedExperiment object in which the object is to be embedded. Must be a superset of rownames(data).

Supported methods

pvalue(x) returns the vector of raw p-values.

log2foldchange(x) returns the vector of log2-fold-change values.

log2average(x) returns the vector of average log2-expression values.

Author

Kevin Rue-Albrecht

Examples

library(limma)
#> 
#> Attaching package: ‘limma’
#> The following object is masked from ‘package:DESeq2’:
#> 
#>     plotMA
#> The following object is masked from ‘package:BiocGenerics’:
#> 
#>     plotMA
library(SummarizedExperiment)

##
# From limma::lmFit() ----
##

sd <- 0.3*sqrt(4/rchisq(100,df=4))
y <- matrix(rnorm(100*6,sd=sd),100,6)
rownames(y) <- paste("Gene",1:100)
y[1:2,4:6] <- y[1:2,4:6] + 2
design <- cbind(Grp1=1, Grp2vs1=c(0,0,0,1,1,1))

fit <- lmFit(y,design)
fit <- eBayes(fit)
tt <- topTable(fit, coef=2)
head(tt)
#>               logFC    AveExpr         t      P.Value    adj.P.Val         B
#> Gene 1    2.0423892  0.9614314  8.433048 3.408431e-06 0.0003408431  4.842099
#> Gene 2    1.6860810  1.0188197  6.749986 2.835182e-05 0.0014175908  2.660428
#> Gene 52  -0.8691408  0.1361762 -3.410451 5.649273e-03 0.1883090998 -2.782987
#> Gene 100  0.7365086  0.1673752  3.132039 9.311872e-03 0.2327968014 -3.284180
#> Gene 60  -0.8395578 -0.4915388 -2.287006 4.252613e-02 0.7535654299 -4.768600
#> Gene 55   0.7314665  0.2451056  2.056950 6.366200e-02 0.7535654299 -5.147974

##
# iSEELimmaResults ----
##

# Simulate the original SummarizedExperiment object
se <- SummarizedExperiment(assays = list(counts=y))

# Package the results in a iSEELimmaResults object
iseede_table <- iSEELimmaResults(tt, row.names=rownames(y))

# Store the iSEELimmaResults object in the SummarizedExperiment rowData
rowData(se)[["iSEEde"]] <- DataFrame(limma=I(iseede_table))

se
#> class: SummarizedExperiment 
#> dim: 100 6 
#> metadata(0):
#> assays(1): counts
#> rownames(100): Gene 1 Gene 2 ... Gene 99 Gene 100
#> rowData names(1): iSEEde
#> colnames: NULL
#> colData names(0):

##
# Methods ----
##

head(pvalue(iseede_table))
#>       Gene 1       Gene 2       Gene 3       Gene 4       Gene 5       Gene 6 
#> 3.408431e-06 2.835182e-05           NA           NA           NA           NA 
head(log2foldchange(iseede_table))
#>   Gene 1   Gene 2   Gene 3   Gene 4   Gene 5   Gene 6 
#> 2.042389 1.686081       NA       NA       NA       NA 
head(log2average(iseede_table))
#>    Gene 1    Gene 2    Gene 3    Gene 4    Gene 5    Gene 6 
#> 0.9614314 1.0188197        NA        NA        NA        NA