The iSEEedgeRResults class

The iSEEedgeRResults class is used to provide an common interface to differential expression results produced by the edgeR package. It provides methods to access common differential expression statistics (e.g., log fold-change, p-value, log2 average abundance).

Details

This class inherits all its slots directly from its parent class DataFrame.

Constructor

iSEEedgeRResults(data, row.names = rownames(data)) creates an instance of a iSEEedgeRResults class, with:

data

A data.frame produced by edgeR::topTags().

row.names

The character vector of rownames for the SummarizedExperiment object in which the object is to be embedded. Must be a superset of rownames(data).

Supported methods

embedContrastResults(x, se, name, ...) embeds x in the column name of rowData(se)[["iSEEde"]].

pValue(x) returns the vector of raw p-values.

log2FoldChange(x) returns the vector of log2-fold-change values.

averageLog2(x) returns the vector of average log2-expression values.

Author

Kevin Rue-Albrecht

Examples

library(edgeR)
#> 
#> Attaching package: ‘edgeR’
#> The following object is masked from ‘package:SingleCellExperiment’:
#> 
#>     cpm
library(SummarizedExperiment)

##
# From edgeR::topTags() ----
##

# generate raw counts from NB, create list object
y <- matrix(rnbinom(80, size = 1, mu = 10), nrow = 20)
d <- DGEList(counts = y, group = rep(1:2, each = 2), lib.size = rep(c(1000:1001), 2))
rownames(d$counts) <- paste("gene", 1:nrow(d$counts), sep = ".")

# estimate common dispersion and find differences in expression
# here we demonstrate the 'exact' methods, but the use of topTags is
# the same for a GLM analysis
d <- estimateCommonDisp(d)
de <- exactTest(d)

# look at top 10
res <- topTags(de)

##
# iSEEedgeRResults ----
##

# Simulate the original SummarizedExperiment object
se <- SummarizedExperiment(assays = list(counts = d$counts))

# Package the results in a iSEEedgeRResults object
iseede_table <- iSEEedgeRResults(res, row.names = rownames(se))

# Store the iSEEedgeRResults object in the SummarizedExperiment rowData
rowData(se)[["iSEEde"]] <- DataFrame(edgeR = I(iseede_table))

se
#> class: SummarizedExperiment 
#> dim: 20 4 
#> metadata(0):
#> assays(1): counts
#> rownames(20): gene.1 gene.2 ... gene.19 gene.20
#> rowData names(1): iSEEde
#> colnames(4): Sample1 Sample2 Sample3 Sample4
#> colData names(0):

##
# Methods ----
##

head(pValue(iseede_table))
#>     gene.1     gene.2     gene.3     gene.4     gene.5     gene.6 
#> 0.00340054 0.19410338         NA 0.39942521         NA 0.29991262 
head(log2FoldChange(iseede_table))
#>    gene.1    gene.2    gene.3    gene.4    gene.5    gene.6 
#> -4.644069 -1.673725        NA  1.286838        NA  1.396489 
head(averageLog2(iseede_table))
#>   gene.1   gene.2   gene.3   gene.4   gene.5   gene.6 
#> 14.00364 14.35156       NA 12.81715       NA 13.65965